pyproject.toml
[project]
name = "rag-pythonclaude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.28.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "google-generativeai>=0.3.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-pinecone>=0.0.3",
    "langchain-text-splitters>=0.0.1",
    "nltk>=3.8",
    "openpyxl>=3.1.0",
    "pinecone-client>=3.0.0",
    "pydantic>=2.0.0",
    "pypdf2>=3.0.0",
    "python-docx>=0.8.11",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "unstructured>=0.10.0",
]
